Fix login not working now that cacheBreaker removed
The issue was that checkAllowedUnauthRoutes logic was wrong, because it considered query string.

Requests for /api/v1/login?cacheBreaker=xyz would return false, and the control flow would fall to the public access policy.

Removing cacheBreaker made checkAllowedUnauthRoutes return true, which caused a higher up if block to execute, which caused an error where a userId was expected but undefined.