Add build-dev script to packages
This differentiates it from the normal build script, which is used for production
releases. Typically we only want to build to a state where a developer can start work, or tests can be run on a package etc.

One direct advantage is that admin panel no longer does a production build after each yyarn install.