Fix CodeShip builds due to missing devops pacakge
Remove the devops package  from .dockerignore. Scripts from this package are
required by CodeShip services containers.

This means the devops package is included in the image, which is
a little redundant but most likely harmless.

Move the Dockerfile out of devops so changes to the Dockerfile don't
invalidate the layer cache unnecessarily.