importSurveys test does not pass a file
In a future commit this will allow the file to be optional, as we will be able to both create and edit a survey without creating/updating survey questions.

This is also not necessary as all of these tests simply check permissions, not anything to do with the survey questions file.

In a future commit we will add a test to cover survey question importing.