MDEV-48 pre fill survey (#4567)
* Remove editSurvey

All validation should be captured in importSurveys, which is used for both creating and editing surveys.

* Minor: Add flag MS1_SYNC_DISABLE

* Survey export use survey code for sheet name

This makes it consistent with survey responses and lets us uniquely identify surveys when importing

* Improve test coverage before refactors

* Survey import by code, one at a time

* Refactor: Split Survey and Survey Question import

* FileInputField can display label

* Add admin panel form field type "file"

* Add clearAllTestData (simpler alternative to clearTestData)

clearnAllTestData simply truncates all tables now that we use tupaia_test instead of the main tupaia db

* importSurveys test does not pass a file

In a future commit this will allow the file to be optional, as we will be able to both create and edit a survey without creating/updating survey questions.

This is also not necessary as all of these tests simply check permissions, not anything to do with the survey questions file.

In a future commit we will add a test to cover survey question importing.

* Add multipartJson middleware

* MDEV-48 Survey CRUD done with CRUD routes instead of surveyImport

Rename importSurveys to SurveyEditor. Mostly the same just a class now with create() and edit() methods. CreateSurvey and EditSurveys call these methods.

Data now comes in like a normal CRUD request (thanks to the multipartJson middleware) rather than on the query string.

* GetSurveys also returns surveyQuestions field.

* MDEV-48 SurveysPage upsert survey questions through normal CRUD forms

* Swap from CRUD with ids to CRUD with names

Limitation of the Admin Panel. Will be cleaned up in RN-910.

* SurveyEditor fixes

- Allow updating a survey without changing data service type
- Allow clearing survey group and permission group
- Other small issues

* Refactor: Move SurveyEditor to apiV2/surveys

* Misc fixes

- Remove duplicate route
- Fix validation

* Fix multipart middleware to handle non-multipart form data as well.

If no file is uploaded, the browser may opt to send the request as Content-Type: application/json instead

* Misc validation fixes

* Revert "Survey export use survey code for sheet name"

This reverts commit f6f0cb06e8e8e47185887e60e15dd8e8376f6cf1.

* Revert to survey questions spreadsheet import with survey name

* Tidy comments

* Update README

* Fix prettier

* (Test issue 2) Add default values to create Survey form

* Simplify Editor/custom fields rendering

* (Test issue 3) Suggest code when creating survey

* (Test issue 4) Hide MS1 field in create survey form

* (Test issues 7 and 9) Fix not being able to blank out fields on survey edit

* (Test issue 6) Fix not being able to change survey code

The implementation before assumed changing survey code would just work unchanged, but I forgot that this was done in the old CRUD form which was deleted, so I had to reimplement the functionality in this new SurveyEditor.

* Minor: preserve stack trace when rethrowing error

* Refactor: move assertCanAddDataElementInGroup

It is specific to survey importing, doesn't belong in utilities.

* (Test issue 10) Update data element configs when data group config changes

* (Test issue 11) Remove check on spreadsheet tab name

* Minor: throw correct error type

* (Test issue 15): Fix typo

* (Test issue 16) Update permission groups helper text

* (Test issue 19) Fix missing import

* (Test issue 20) New data elements use existing permission group

---------

Co-authored-by: Andrew <vanbeekandrew@gmail.com>